% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_calculations.R
\name{get_number_per_region_per_1000_residents_plus_buffer}
\alias{get_number_per_region_per_1000_residents_plus_buffer}
\title{Count points per 1000 residents in regions plus a buffer}
\usage{
get_number_per_region_per_1000_residents_plus_buffer(
  region_shp,
  point_data,
  pop_data,
  region_id_col = "ONS_ID",
  pop_col = "SF_TotalPop",
  crs = 32189,
  buffer_m = 50
)
}
\arguments{
\item{region_shp}{Sf object with polygonal shapes, one row per region, with at
least one column providing unique region identifiers.}

\item{point_data}{Sf object with point data to aggregate to regions.}

\item{pop_data}{Data frame giving regional population data, including at least
one column with unique region identifiers and one column with population data.}

\item{region_id_col}{Character, column name in `region_shp` containing unique
region identifiers.}

\item{pop_col}{character, column name in `pop_data` containing populations.}

\item{crs}{The CRS to convert regions and points to. Defaults to 32189
(NAD83 / MTM zone 9) for Ottawa, Ontario.}

\item{buffer_m}{Buffer in meters to apply to regions.}
}
\value{
A tibble with one row per region and the number of points within each.
}
\description{
Count points per 1000 residents in regions plus a buffer
}
