% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_calculations.R
\name{get_number_per_region_plus_buffer}
\alias{get_number_per_region_plus_buffer}
\title{Count points in regions plus a buffer}
\usage{
get_number_per_region_plus_buffer(
  region_shp,
  point_data,
  region_id_col = "ONS_ID",
  crs = 32189,
  buffer_m = 50
)
}
\arguments{
\item{region_shp}{Sf object with polygonal shapes, one row per region, with at
least one column providing unique region identifiers.}

\item{point_data}{Sf object with point data to aggregate to regions.}

\item{region_id_col}{Character, column name in `region_shp` containing unique
region identifiers.}

\item{crs}{The CRS to convert regions and points to. Defaults to 32189
(NAD83 / MTM zone 9) for Ottawa, Ontario.}

\item{buffer_m}{Buffer in meters to apply to regions.}
}
\value{
A tibble with one row per region and the number of points within each.
}
\description{
Count points in regions plus a buffer
}
