% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_calculations.R
\name{get_pct_within_15_mins}
\alias{get_pct_within_15_mins}
\title{Use precomputed distance tables to get the percentage of residents within 15-minute walk of features}
\usage{
get_pct_within_15_mins(
  od_table,
  from_id_col = "DBUID",
  to_id_col,
  froms_to_ons_sli,
  dbpops
)
}
\arguments{
\item{od_table}{tibble output from `valhallr::od_table()` with columns `distance` and `time`}

\item{from_id_col}{Character, name of column with unique origin identifiers. Default "DBUID".}

\item{to_id_col}{Character, name of column with unique destination identifiers.}

\item{froms_to_ons_sli}{Single-link indicator (SLI) from origins to ONS neighbourhoods. Consider using `neighbourhoodstudy::sli_das_gen3_mape`.}

\item{dbpops}{Tibble, in the format of (or consider using) `neighbourhoodstudy::ottawa_dbs_pop2021`.}
}
\value{
Tibble with percent of residents within 15-minute walk at ONS neighbourhood and Ottawa levels.
}
\description{
Use precomputed distance tables to get the percentage of residents within 15-minute walk of features
}
