% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_gen2_shp}
\alias{ons_gen2_shp}
\title{Ottawa Neighbourhood Study Spatial Boundaries (Gen2)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 111 rows and 9 columns.
}
\usage{
ons_gen2_shp
}
\description{
The ONS Neighbourhood Boundaries were created by the Ottawa Neighbourhood Study
(ONS) to analyse population statistics and are not indicative of actual neighbourhood limits.
}
\details{
Available publicly from the City of Ottawa: \url{https://open.ottawa.ca/datasets/32fe76b71c5e424fab19fec1f180ec18_0/explore}
}
\seealso{
ons_shp_gen3

Other shapefiles: 
\code{\link{ons_gen3_shp}},
\code{\link{ottawa_das_shp2021}},
\code{\link{ottawa_dbs_shp2021}},
\code{\link{ottawa_phhs}}
}
\concept{shapefiles}
\keyword{datasets}
