% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_gen3_pop2021}
\alias{ons_gen3_pop2021}
\title{Statistics Canada population statistics for ONS Gen3 Neighbourhoods}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161 rows and 8 columns.
}
\usage{
ons_gen3_pop2021
}
\description{
SF_TotalPop	Total population, non-institutional and institutional residents, 100% Census data.
SF_NIPOP	Total non-institutional population in private households, 100% Census data.
SF_PTDWELL	Total number of private dwellings, 100% Census data.
SF_PODWELL_UR	Total number of private dwellings occupied by usual residents, 100% Census data.
LF_POP_PRIV	Total population in private households from long form, 25% Census data.
LF_Households	Total number of households, 25% Census data.
}
\seealso{
ons_shp_gen3
}
\keyword{datasets}
