% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_gen3_shp}
\alias{ons_gen3_shp}
\title{Ottawa Neighbourhood Study Spatial Boundaries (Gen3)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 161 rows and 7 columns.
}
\usage{
ons_gen3_shp
}
\description{
The ONS Neighbourhood Boundaries were created by the Ottawa Neighbourhood Study
(ONS) to analyse population statistics and are not indicative of actual neighbourhood limits.
}
\seealso{
ons_shp_gen2

Other shapefiles: 
\code{\link{ons_gen2_shp}},
\code{\link{ottawa_das_shp2021}},
\code{\link{ottawa_dbs_shp2021}},
\code{\link{ottawa_phhs_shp}}
}
\concept{shapefiles}
\keyword{datasets}
