% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_shp_gen2}
\alias{ons_shp_gen2}
\title{Ottawa Neighbourhood Study Spatial Boundaries (Gen2)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 111 rows and 9 columns.
}
\usage{
ons_shp_gen2
}
\description{
Ottawa Neighbourhood Study Spatial Boundaries (Gen2)
}
\keyword{datasets}
