% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ottawa_dbs_pop2021}
\alias{ottawa_dbs_pop2021}
\title{Statistics Canada populations for 2021 Dissemination Blocks (DBs) in Ottawa}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8559 rows and 5 columns.
}
\usage{
ottawa_dbs_pop2021
}
\description{
Data taken from the Geographic Attribute File
https://www12.statcan.gc.ca/census-recensement/2021/geo/aip-pia/attribute-attribs/index2021-eng.cfm
}
\details{
Column names have been shortened for ease of use. These definitions are from the
record layout referenced below:

dbpop21:	Number (9)	2021 Census dissemination block population.
dbtdwell2021:	Number (9)	2021 Census dissemination block total private dwellings.
dburbdwell2021: 	Number (9)	2021 Census private dwellings occupied by usual residents.
dbarea2021: 	Number (14.4)	2021 Census dissemination block land area.

For record layout: https://www150.statcan.gc.ca/n1/pub/92-151-g/2021001/tbl/tbl4_1-eng.htm
}
\seealso{
ottawa_dbs_shp
}
\keyword{datasets}
