% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ottawa_dbs_shp2021}
\alias{ottawa_dbs_shp2021}
\title{2021 Statistics Canada Dissemination Blocks (DBs) in Ottawa}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8559 rows and 7 columns.
}
\usage{
ottawa_dbs_shp2021
}
\description{
https://www12.statcan.gc.ca/census-recensement/2021/geo/sip-pis/boundary-limites/index2021-eng.cfm?year=21
}
\seealso{
Other shapefiles: 
\code{\link{ons_gen2_shp}},
\code{\link{ons_gen3_shp}},
\code{\link{ottawa_das_shp2021}},
\code{\link{ottawa_phhs}}
}
\concept{shapefiles}
\keyword{datasets}
