% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ottawa_phhs}
\alias{ottawa_phhs}
\title{Pseudo-Households in Ottawa}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 124891 rows and 13 columns.
}
\usage{
ottawa_phhs
}
\description{
The following description is from the Government of Canada:
https://open.canada.ca/data/en/dataset/b3a1d603-19ca-466c-ae95-b5185e56addf
}
\details{
"The Pseudo-Household Demographic Distribution is a geospatial representative
distribution of demographic data (population and households) derived from the
Canadian Census from Statistics Canada. Demography is distributed within
Dissemination Blocks along roadways, providing a more accurate geospatial
distribution while still aligning with published Census geographies.
Pseudo-household demographics are currently used to calculate broadband
Internet service availability, but are equally applicable to other
disciplines requiring a spatial distribution of households or population."
}
\seealso{
Other shapefiles: 
\code{\link{ons_shp_gen2}},
\code{\link{ons_shp_gen3}},
\code{\link{ottawa_shp_dbs}}
}
\concept{shapefiles}
\keyword{datasets}
