% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{query_osm_api}
\alias{query_osm_api}
\title{Query the OpenStreetMaps (OSM) Overpass API for Amenities and Shops}
\usage{
query_osm_api(
  bounding_box_shp,
  amenities = "",
  shops = "",
  process = TRUE,
  drop_unnamed_elements = TRUE
)
}
\arguments{
\item{bounding_box_shp}{An sf object that will be used to create a square bounding box for the search.}

\item{amenities}{Optional. A character vector of amenity key values to search for.}

\item{shops}{Optional. A character vector of shop key values to search for.}

\item{process}{Boolean, default TRUE. Processes data to return a tidy result. If FALSE, returns raw Overpass API response.}

\item{drop_unnamed_elements}{Boolean, default TRUE. Removes elements with no name that seem to be returned in error when searching e.g. for cafes.}
}
\value{
A tibble of OSM elements with point locations.
}
\description{
Retrieve data from OSM for any number of amenity or shop key values and
return a tidy list of point-wise results. Polygons are converted to centroids.
}
