% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sli_das_gen3_mape}
\alias{sli_das_gen3_mape}
\title{Single-Link Indicator (SLI) from 2021 StatsCan DAs to Gen3 Neighbourhoods}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1392 rows and 2 columns.
}
\usage{
sli_das_gen3_mape
}
\description{
Created using package `sliopt` and optimizing for mean average percentage error (MAPE).
}
\seealso{
Other slis: 
\code{\link{sli_dbs_gen3_maxoverlap}}
}
\concept{slis}
\keyword{datasets}
