% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sli_dbs_gen3_maxoverlap}
\alias{sli_dbs_gen3_maxoverlap}
\title{Max-Area Single-Link Indicator (SLI) from 2021 StatsCan DBs to Gen3 Neighbourhoods}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8559 rows and 2 columns.
}
\usage{
sli_dbs_gen3_maxoverlap
}
\description{
Dissemination Blocks are linked to the single ONS Gen3 neighbourhood they most
overlap.
}
\concept{slis}
\keyword{datasets}
