% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ons_shp_gen3}
\alias{ons_shp_gen3}
\title{Ottawa Neighbourhood Study Spatial Boundaries (Gen3)}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 162 rows and 7 columns.
}
\usage{
ons_shp_gen3
}
\description{
Ottawa Neighbourhood Study Spatial Boundaries (Gen3)
}
\keyword{datasets}
